refactor: split codegen optimizer tests
