+++ tests/codegen/optimizer/constant_folding.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+#[path = "constant_folding/expressions.rs"]
+mod expressions;
+#[path = "constant_folding/pruning.rs"]
+mod pruning;
