+++ tests/codegen/optimizer/constant_propagation.rs
@@ -1,0 +1,10 @@
+use super::*;
+
+#[path = "constant_propagation/straight_line.rs"]
+mod straight_line;
+#[path = "constant_propagation/control_paths.rs"]
+mod control_paths;
+#[path = "constant_propagation/collections.rs"]
+mod collections;
+#[path = "constant_propagation/loops.rs"]
+mod loops;
