+++ tests/codegen/optimizer/constant_propagation/loops.rs
@@ -1,0 +1,10 @@
+use super::*;
+
+#[path = "loops/for_loops.rs"]
+mod for_loops;
+#[path = "loops/while_loops.rs"]
+mod while_loops;
+#[path = "loops/foreach_loops.rs"]
+mod foreach_loops;
+#[path = "loops/loop_state.rs"]
+mod loop_state;
