+++ tests/codegen/optimizer/dead_code_elimination.rs
@@ -1,0 +1,14 @@
+use super::*;
+
+#[path = "dead_code_elimination/basics.rs"]
+mod basics;
+#[path = "dead_code_elimination/tries.rs"]
+mod tries;
+#[path = "dead_code_elimination/switches.rs"]
+mod switches;
+#[path = "dead_code_elimination/guards.rs"]
+mod guards;
+#[path = "dead_code_elimination/tail_sinking.rs"]
+mod tail_sinking;
+#[path = "dead_code_elimination/normalization.rs"]
+mod normalization;
