+++ tests/codegen/optimizer/dead_code_elimination/guards.rs
@@ -1,0 +1,8 @@
+use super::*;
+
+#[path = "guards/outer_guards.rs"]
+mod outer_guards;
+#[path = "guards/excluded_guards.rs"]
+mod excluded_guards;
+#[path = "guards/composite_guards.rs"]
+mod composite_guards;
