+++ tests/codegen/optimizer/dead_code_elimination/normalization.rs
@@ -1,0 +1,101 @@
+use super::*;
+
+#[test]
+fn test_dead_code_elimination_collapses_identical_if_branches() {
+    let out = compile_and_run(
+        r#"<?php
+function step($label, $ret) {
+    echo $label;
+    return $ret;
+}
+if (step("c", false)) {
+    echo "X";
+} else {
+    echo "X";
+}
+"#,
+    );
+
+    assert_eq!(out, "cX");
+}
+
+#[test]
+fn test_dead_code_elimination_merges_identical_if_chain_bodies_with_short_circuit() {
+    let out = compile_and_run(
+        r#"<?php
+function step($label, $ret) {
+    echo $label;
+    return $ret;
+}
+if (step("a", true)) {
+    echo "X";
+} else {
+    if (step("b", true)) {
+        echo "X";
+    } else {
+        echo "Y";
+    }
+}
+"#,
+    );
+
+    assert_eq!(out, "aX");
+}
+
+#[test]
+fn test_dead_code_elimination_recursively_merges_longer_if_chains() {
+    let out = compile_and_run(
+        r#"<?php
+function step($label, $ret) {
+    echo $label;
+    return $ret;
+}
+if (step("a", false)) {
+    echo "X";
+} else {
+    if (step("b", false)) {
+        echo "X";
+    } else {
+        if (step("c", true)) {
+            echo "X";
+        } else {
+            echo "Y";
+        }
+    }
+}
+echo "|";
+if (step("d", false)) {
+    echo "X";
+} else {
+    if (step("e", false)) {
+        echo "Y";
+    } else {
+        if (step("f", true)) {
+            echo "Y";
+        } else {
+            echo "X";
+        }
+    }
+}
+"#,
+    );
+
+    assert_eq!(out, "abcX|defY");
+}
+
+#[test]
+fn test_dead_code_elimination_flattens_nested_single_path_ifs() {
+    let out = compile_and_run(
+        r#"<?php
+$a = true;
+$b = true;
+if ($a) {
+    if ($b) {
+        echo 7;
+    }
+}
+"#,
+    );
+
+    assert_eq!(out, "7");
+}
