+++ tests/codegen/optimizer/dead_code_elimination/switches.rs
@@ -1,0 +1,12 @@
+use super::*;
+
+#[path = "switches/case_shadowing.rs"]
+mod case_shadowing;
+#[path = "switches/guarded_cases.rs"]
+mod guarded_cases;
+#[path = "switches/exhaustive_suffixes.rs"]
+mod exhaustive_suffixes;
+#[path = "switches/normalization.rs"]
+mod normalization;
+#[path = "switches/tail_paths.rs"]
+mod tail_paths;
