+++ tests/codegen/optimizer/dead_code_elimination/tries.rs
@@ -1,0 +1,12 @@
+use super::*;
+
+#[path = "tries/try_pruning.rs"]
+mod try_pruning;
+#[path = "tries/catch_pruning.rs"]
+mod catch_pruning;
+#[path = "tries/finally_paths.rs"]
+mod finally_paths;
+#[path = "tries/try_inlining.rs"]
+mod try_inlining;
+#[path = "tries/tail_paths.rs"]
+mod tail_paths;
