+++ tests/codegen/optimizer/dead_code_elimination/tries/try_inlining.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+#[path = "try_inlining/pure_calls.rs"]
+mod pure_calls;
+#[path = "try_inlining/callable_aliases.rs"]
+mod callable_aliases;
